Created base classes and tests for books module

+++ src/main/java/com/renanmartins/bookstoremanager/books/dto/BookResponseDTO.java
@@ -16,9 +16,9 @@ public class BookResponseDTO {
 
     private String isbn;
 
-    private Long pages;
+    private Integer pages;
 
-    private Long chapters;
+    private Integer chapters;
 
     private AuthorDTO author;
